requirements.txt
blinker==1.6.2
certifi==2023.5.7
chardet==3.0.4
click==8.1.3
Flask==2.3.2
Flask-GoogleCharts==0.1.0
Flask-WTF==1.1.1
future==0.18.3
idna==2.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
lib==4.0.0
MarkupSafe==2.1.3
paho-mqtt==1.4.0
pytz==2018.5
requests==2.20.1
requests-toolbelt==1.0.0
tzlocal==2.0.0
urllib3==1.24.3
webexteamssdk==1.1.1
Werkzeug==2.3.6
WTForms==2.2.1
zipp==3.15.0